chromium/third_party/blink/web_tests/http/tests/fetch/chromium/use-counters.html

<!doctype html>
<head>
<meta name="timeout" content="long"></meta>
<script src = "/resources/get-host-info.js?pipe=sub"></script>
<script src = "/resources/testharness.js"></script>
<script src = "/resources/testharnessreport.js"></script>
<script src = "/serviceworker/resources/test-helpers.js"></script>
<script src = "/fetch/resources/fetch-test-options.js"></script>
<script src = "/fetch/resources/fetch-test-helpers.js"></script>
</head>
<body>
<script>
// The tests in this file are solely used for for the UseCounters associated
// with crbug.com/1072350.
const {BASE_ORIGIN, OTHER_ORIGIN} = get_fetch_test_options();

const kFetchAPINonGetOrHeadOpaqueResponse = 3316,
      kFetchAPINonGetOrHeadOpaqueResponseWithRedirect = 3317;

const cross_origin_url = OTHER_ORIGIN + '/fetch/resources/fetch-status.php?status=200';

// These test that the UseCounters are not incorrectly counted.
// No conditions met: (GET, "cors", basic)
promise_test(t => {
  return fetch("/fetch/resources/fetch-status.php?status=200",
               {method: 'GET', mode: 'cors'})
    .then(response => {
      assert_equals(response.type, 'basic');
      assert_false(internals.isUseCounted(document,
        kFetchAPINonGetOrHeadOpaqueResponse));
      assert_false(internals.isUseCounted(document,
        kFetchAPINonGetOrHeadOpaqueResponseWithRedirect));
    });
}, "Requests whose method=GET, mode=cors, and response type=basic do not " +
   "count the UseCounters");

// One condition met: (POST, "cors" basic)
promise_test(t => {
  return fetch("/fetch/resources/fetch-status.php?status=200",
               {method: 'POST', mode: 'cors'})
    .then(response => {
      assert_equals(response.type, 'basic');
      assert_false(internals.isUseCounted(document,
        kFetchAPINonGetOrHeadOpaqueResponse));
      assert_false(internals.isUseCounted(document,
        kFetchAPINonGetOrHeadOpaqueResponseWithRedirect));
    });
}, "Requests whose method=POST, mode=cors, and response type=basic do not " +
   "count the UseCounters");

// One condition met: (GET, "no-cors" basic)
promise_test(t => {
  return fetch("/fetch/resources/fetch-status.php?status=200",
               {method: 'GET', mode: 'no-cors'})
    .then(response => {
      assert_equals(response.type, 'basic');
      assert_false(internals.isUseCounted(document,
        kFetchAPINonGetOrHeadOpaqueResponse));
      assert_false(internals.isUseCounted(document,
        kFetchAPINonGetOrHeadOpaqueResponseWithRedirect));
    });
}, "Requests whose method=GET, mode=no-cors, and response type=basic do not " +
   "count the UseCounters");

// One condition met: (GET, "no-cors", opaqueredirect) - {redirect: 'manual'}
promise_test(t => {
  const redirect_target_url =
    BASE_ORIGIN + '/fetch/resources/fetch-status.php?status=200';
  const redirect_original_url =
    BASE_ORIGIN + '/serviceworker/resources/redirect.php?Redirect=' +
    redirect_target_url;

  return fetch(redirect_original_url,
               {method: 'GET', mode: 'no-cors', redirect: 'manual'})
    .then(response => {
      assert_equals(response.type, 'opaqueredirect');
      assert_false(internals.isUseCounted(document,
        kFetchAPINonGetOrHeadOpaqueResponse));
      assert_false(internals.isUseCounted(document,
        kFetchAPINonGetOrHeadOpaqueResponseWithRedirect));
    });
}, "Requests whose method=GET, mode=no-cors, and response " +
   "type=opaqueredirect do not count the UseCounters");

// Two conditions met: (POST, "no-cors", basic)
promise_test(t => {
  return fetch('/fetch/resources/fetch-status.php?status=200',
               {method: 'POST', mode: 'no-cors'})
    .then(response => {
      assert_equals(response.type, 'basic');
      assert_false(internals.isUseCounted(document,
        kFetchAPINonGetOrHeadOpaqueResponse));
      assert_false(internals.isUseCounted(document,
        kFetchAPINonGetOrHeadOpaqueResponseWithRedirect));
    });
}, "Requests whose method=POST, mode=no-cors, and response type=basic do not " +
   "count the UseCounters");

// Two conditions met: (GET, "no-cors", opaque)
promise_test(t => {
  return fetch(cross_origin_url, {method: 'GET', mode: 'no-cors'})
    .then(response => {
      assert_equals(response.type, 'opaque');
      assert_false(internals.isUseCounted(document,
        kFetchAPINonGetOrHeadOpaqueResponse));
      assert_false(internals.isUseCounted(document,
        kFetchAPINonGetOrHeadOpaqueResponseWithRedirect));
    });
}, "Requests whose method=GET, mode=no-cors, and response type=opaque do not " +
   "count the UseCounters");

// These test that the UseCounters are being correctly counted.

promise_test(t => {
  return fetch(cross_origin_url, {method: 'POST', mode: 'no-cors'})
    .then(response => {
      assert_equals(response.type, 'opaque');
      assert_true(internals.isUseCounted(document,
        kFetchAPINonGetOrHeadOpaqueResponse));
      assert_false(internals.isUseCounted(document,
        kFetchAPINonGetOrHeadOpaqueResponseWithRedirect));

      internals.clearUseCounter(document, kFetchAPINonGetOrHeadOpaqueResponse);
    });
}, "Requests whose method=POST, mode=no-cors, and response type=opaque " +
   "counts one of the UseCounters");

promise_test(t => {
  const redirect_target_url =
    OTHER_ORIGIN + '/fetch/resources/fetch-status.php?status=200';
  const redirect_original_url =
    OTHER_ORIGIN + '/serviceworker/resources/redirect.php?Redirect=' +
    redirect_target_url;

  return fetch(redirect_original_url, {method: 'POST', mode: 'no-cors'})
    .then(response => {
      assert_equals(response.type, 'opaque');
      assert_true(internals.isUseCounted(document,
        kFetchAPINonGetOrHeadOpaqueResponse));
      assert_true(internals.isUseCounted(document,
        kFetchAPINonGetOrHeadOpaqueResponseWithRedirect));

      internals.clearUseCounter(document,
        kFetchAPINonGetOrHeadOpaqueResponse);
      internals.clearUseCounter(document,
        kFetchAPINonGetOrHeadOpaqueResponseWithRedirect);
    });
}, "Requests whose method=POST, mode=no-cors, and response type=opaque an " +
   "experiences a cross-origin => cross-origin redirect counts both UseCounters");

promise_test(t => {
  const redirect_target_url =
    OTHER_ORIGIN + '/fetch/resources/fetch-status.php?status=200';
  const redirect_original_url =
    OTHER_ORIGIN + '/serviceworker/resources/redirect.php?Redirect=' +
    redirect_target_url;

  return fetch(redirect_original_url, {method: 'POST', mode: 'no-cors'})
    .then(response => {
      assert_equals(response.type, 'opaque');
      assert_true(internals.isUseCounted(document,
        kFetchAPINonGetOrHeadOpaqueResponse));
      assert_true(internals.isUseCounted(document,
        kFetchAPINonGetOrHeadOpaqueResponseWithRedirect));

      internals.clearUseCounter(document,
        kFetchAPINonGetOrHeadOpaqueResponse);
      internals.clearUseCounter(document,
        kFetchAPINonGetOrHeadOpaqueResponseWithRedirect);
    });
}, "Requests whose method=POST, mode=no-cors, and response type=opaque an " +
   "experiences a same-origin => cross-origin redirect counts both UseCounters");

done();

</script>
</body>
